fix: Added new structure to Linkedin icon
