+++ src/components/index.ts
@@ -148,7 +148,9 @@ export * from './social-icons/Medium';
 export * from './social-icons/Reddit';
 export * from './social-icons/Twitter';
 export * from './social-icons/Facebook';
-export * from './social-icons/Linkedin';
+export * from './social-icons/Linkedin/Linkedin';
+export * from './social-icons/Linkedin/Linkedin18';
+export * from './social-icons/Linkedin/Linkedin24';
 export * from './social-icons/Instagram/Instagram';
 export * from './social-icons/Instagram/Instagram18';
 export * from './social-icons/Instagram/Instagram24';
