--- src/components/social-icons/Linkedin.tsx
@@ -1,62 +1,0 @@
-import React, { FC } from 'react';
-import { ISocialIconProps } from './type';
-
-export const IconLinkedin: FC<ISocialIconProps> = ({
-	size = 16,
-	color = 'currentColor',
-}) => {
-	const size18 = (
-		<svg
-			width='18'
-			height='18'
-			viewBox='0 0 18 18'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M12 6C13.1935 6 14.3381 6.47411 15.182 7.31802C16.0259 8.16193 16.5 9.30653 16.5 10.5V14.75C16.5 15.3023 16.0523 15.75 15.5 15.75H14.5C13.9477 15.75 13.5 15.3023 13.5 14.75V10.5C13.5 10.1022 13.342 9.72064 13.0607 9.43934C12.7794 9.15804 12.3978 9 12 9C11.6022 9 11.2206 9.15804 10.9393 9.43934C10.658 9.72064 10.5 10.1022 10.5 10.5V14.75C10.5 15.3023 10.0523 15.75 9.5 15.75H8.5C7.94772 15.75 7.5 15.3023 7.5 14.75V10.5C7.5 9.30653 7.97411 8.16193 8.81802 7.31802C9.66193 6.47411 10.8065 6 12 6Z'
-				fill={color}
-			/>
-			<path
-				d='M3.5 6.75H2.5C1.94772 6.75 1.5 7.19772 1.5 7.75V14.75C1.5 15.3023 1.94772 15.75 2.5 15.75H3.5C4.05228 15.75 4.5 15.3023 4.5 14.75V7.75C4.5 7.19772 4.05228 6.75 3.5 6.75Z'
-				fill={color}
-			/>
-			<path
-				d='M3 4.5C3.82843 4.5 4.5 3.82843 4.5 3C4.5 2.17157 3.82843 1.5 3 1.5C2.17157 1.5 1.5 2.17157 1.5 3C1.5 3.82843 2.17157 4.5 3 4.5Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	const size24 = (
-		<svg
-			width='24'
-			height='24'
-			viewBox='0 0 24 24'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M16 8C17.5913 8 19.1174 8.63214 20.2426 9.75736C21.3679 10.8826 22 12.4087 22 14V20C22 20.5523 21.5523 21 21 21H19C18.4477 21 18 20.5523 18 20V14C18 13.4696 17.7893 12.9609 17.4142 12.5858C17.0391 12.2107 16.5304 12 16 12C15.4696 12 14.9609 12.2107 14.5858 12.5858C14.2107 12.9609 14 13.4696 14 14V20C14 20.5523 13.5523 21 13 21H11C10.4477 21 10 20.5523 10 20V14C10 12.4087 10.6321 10.8826 11.7574 9.75736C12.8826 8.63214 14.4087 8 16 8Z'
-				fill={color}
-			/>
-			<path
-				d='M5 9H3C2.44772 9 2 9.44772 2 10V20C2 20.5523 2.44772 21 3 21H5C5.55228 21 6 20.5523 6 20V10C6 9.44772 5.55228 9 5 9Z'
-				fill={color}
-			/>
-			<path
-				d='M4 6C5.10457 6 6 5.10457 6 4C6 2.89543 5.10457 2 4 2C2.89543 2 2 2.89543 2 4C2 5.10457 2.89543 6 4 6Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	switch (size.toString()) {
-		case '16':
-			return size18;
-		case '24':
-			return size24;
-		default:
-			return size18;
-	}
-};
