+++ src/components/social-icons/Linkedin/Linkedin18.tsx
@@ -1,0 +1,28 @@
+import React, { FC } from 'react';
+import { ISocialIconProps } from '../type';
+
+export const IconLinkedin18: FC<ISocialIconProps> = ({
+	size = 18,
+	color = 'currentColor',
+}) => (
+	<svg
+		width={size}
+		height={size}
+		viewBox='0 0 18 18'
+		fill='none'
+		xmlns='http://www.w3.org/2000/svg'
+	>
+		<path
+			d='M12 6C13.1935 6 14.3381 6.47411 15.182 7.31802C16.0259 8.16193 16.5 9.30653 16.5 10.5V14.75C16.5 15.3023 16.0523 15.75 15.5 15.75H14.5C13.9477 15.75 13.5 15.3023 13.5 14.75V10.5C13.5 10.1022 13.342 9.72064 13.0607 9.43934C12.7794 9.15804 12.3978 9 12 9C11.6022 9 11.2206 9.15804 10.9393 9.43934C10.658 9.72064 10.5 10.1022 10.5 10.5V14.75C10.5 15.3023 10.0523 15.75 9.5 15.75H8.5C7.94772 15.75 7.5 15.3023 7.5 14.75V10.5C7.5 9.30653 7.97411 8.16193 8.81802 7.31802C9.66193 6.47411 10.8065 6 12 6Z'
+			fill={color}
+		/>
+		<path
+			d='M3.5 6.75H2.5C1.94772 6.75 1.5 7.19772 1.5 7.75V14.75C1.5 15.3023 1.94772 15.75 2.5 15.75H3.5C4.05228 15.75 4.5 15.3023 4.5 14.75V7.75C4.5 7.19772 4.05228 6.75 3.5 6.75Z'
+			fill={color}
+		/>
+		<path
+			d='M3 4.5C3.82843 4.5 4.5 3.82843 4.5 3C4.5 2.17157 3.82843 1.5 3 1.5C2.17157 1.5 1.5 2.17157 1.5 3C1.5 3.82843 2.17157 4.5 3 4.5Z'
+			fill={color}
+		/>
+	</svg>
+);
